feat(backend): Add and organize folders structure
