+++ backend/src/modules/auth/repositories/prisma-auth.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/modules/prisma/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { hash } from 'bcrypt';
 import { RefreshTokenMapper } from '../mappers/auth.mapper';
 import { AuthRepository as AuthRepositoryInterface } from './auth.repository';
