+++ backend/src/modules/engagement/engagement.module.ts
@@ -1,0 +1,4 @@
+import { Module } from '@nestjs/common';
+
+@Module({})
+export class EngagementModule {}
