+++ backend/src/modules/user/repositories/prisma/prisma-user-provider.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/modules/prisma/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 import { UserProviderRepository } from '../user-provider.repository';
 import { ProviderType } from '@prisma/client';
 import { UserMapper } from '../../mappers/user.mapper';
