+++ backend/src/modules/user/user.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { PrismaModule } from '../prisma/prisma.module';
+import { PrismaModule } from '../../prisma/prisma.module';
 import { AuthModule } from '../auth/auth.module';
 import { UserRepository } from './repositories/user.repository';
 import { PrismaRepository } from './repositories/prisma/prisma-user.repository';
